Move CustomerioState to state package
