+++ src/main/java/com/hedvig/notificationService/customerio/CustomerioEventCreator.kt
@@ -1,5 +1,7 @@
 package com.hedvig.notificationService.customerio
 
+import com.hedvig.notificationService.customerio.state.CustomerioState
+
 interface
 CustomerioEventCreator {
     fun createTmpSignedInsuranceEvent(customerioState: CustomerioState): Map<String, Any?>
