+++ src/test/java/com/hedvig/notificationService/customerio/CustomerioServiceContructionTest.kt
@@ -1,7 +1,7 @@
 package com.hedvig.notificationService.customerio
 
 import com.hedvig.customerio.CustomerioClient
-import com.hedvig.notificationService.customerio.repository.InMemoryCustomerIOStateRepository
+import com.hedvig.notificationService.customerio.state.InMemoryCustomerIOStateRepository
 import io.mockk.MockKAnnotations
 import io.mockk.impl.annotations.MockK
 import org.junit.Before
