+++ src/test/java/com/hedvig/notificationService/customerio/CustomerioServicePostEventTest.kt
@@ -1,7 +1,7 @@
 package com.hedvig.notificationService.customerio
 
 import com.hedvig.customerio.CustomerioClient
-import com.hedvig.notificationService.customerio.repository.InMemoryCustomerIOStateRepository
+import com.hedvig.notificationService.customerio.state.InMemoryCustomerIOStateRepository
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
